8.score/requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
beautifulsoup4
pymupdf
arxiv
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
langchain-huggingface
tqdm
faiss-cpu
langchain-chroma
streamlit
duckdb
langchain-groq
duckduckgo-search
SQLAlchemy
unstructured
youtube-transcript-api
validators
numexpr
pytube
openai
langchainhub
google-api-python-client
google-auth-oauthlib
google-auth-httplib2